requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
colorlog>=6.7.0
streamlit>=1.28.0
plotly>=5.17.0
python-telegram-bot>=20.0
pandas-ta>=0.3.14
psutil>=5.9.0